chore: Refactor BE file structure and DB connection.
