--- .env
@@ -1,2 +1,0 @@
-DATABASE=mongodb://devUser:HerokuApp123#@ds161751.mlab.com:61751/freeyourstuff
-
